@@ -1,10 +1,21 @@
 {
-  "name": "starter-react-app",
+  "name": "recipe-app",
   "version": "0.1.0",
   "private": true,
+  "dependencies": {
+    "@heroicons/react": "^1.0.6",
+    "@splidejs/react-splide": "^0.7.12",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.11.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
+  },
   "scripts": {
-    "start": "node server.js",
-    "dev": "react-scripts start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -27,16 +38,7 @@
       "last 1 safari version"
     ]
   },
-  "dependencies": {
-    "express": "^4.17.3"
-  },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
-    "web-vitals": "^2.1.4"
+    "tailwindcss": "^3.3.3"
   }
 }
